[Mono.Android] Share Java interface-hierarchy walk between value/type managers
Extract the Class.getInterfaces() recursive walk into a shared
JavaInterfaceHierarchy.FindFirst helper and use it from both
CoreClrJavaMarshalValueManager.CreatePeerInstance and
TrimmableTypeMap.TryMatchInterfaces, removing the duplicated JNI plumbing
(getInterfaces method id + traversal) from both.

No behavior change; consolidates the interface-derived-proxy resolution
added in the previous commit.
